+++ zokrates_cli/Cargo.toml
@@ -22,6 +22,8 @@ zokrates_core = { version = "0.5", path = "../zokrates_core", default-features =
 zokrates_fs_resolver = { version = "0.5", path = "../zokrates_fs_resolver"}
 serde_json = "1.0"
 dirs = "3.0.1"
+log = "0.4.11"
+lazy_static = "1.4.0"
 
 [dev-dependencies]
 glob = "0.2.11"
