+++ zokrates_cli/src/ops/mod.rs
@@ -1,0 +1,12 @@
+pub mod check;
+pub mod compile;
+pub mod compute_witness;
+#[cfg(any(feature = "bellman", feature = "ark", feature = "libsnark"))]
+pub mod export_verifier;
+#[cfg(any(feature = "bellman", feature = "ark", feature = "libsnark"))]
+pub mod generate_proof;
+pub mod print_proof;
+#[cfg(any(feature = "bellman", feature = "ark", feature = "libsnark"))]
+pub mod setup;
+#[cfg(any(feature = "bellman", feature = "ark", feature = "libsnark"))]
+pub mod verify;
