+++ zokrates_core/src/proof_system/scheme/mod.rs
@@ -6,6 +6,10 @@ pub mod gm17;
 pub mod groth16;
 pub mod pghr13;
 
+pub use self::gm17::GM17;
+pub use self::groth16::G16;
+pub use self::pghr13::PGHR13;
+
 pub trait Scheme<T: Field> {
     type VerificationKey: Serialize + DeserializeOwned;
     type ProofPoints: Serialize + DeserializeOwned;
